fix profile (T.Yui)
git-svn-id: http://amethyst.yui.ne.jp/svn/pc98/np2/trunk@705 7c2116fe-53d0-4aec-ab59-80607f05abc0

+++ win9x/ini.cpp
@@ -246,7 +246,7 @@ const INITBL	*pterm;
 				GetPrivateProfileString(title, p->item, work,
 												work, sizeof(work), path);
 				val = (UINT32)milstr_solveHEX(work);
-				*(UINT16 *)p->value = (UINT32)val;
+				*(UINT32 *)p->value = (UINT32)val;
 				break;
 
 			case INITYPE_BYTE3:
